refactor(app): batch1 — extract useAppBootstrap + AppProviders, shrink MainApp
